refactor(fe): [Issue #101-2] ReceiptImageCropModal の画像操作を hook に分離
useImageCropSelection に PanResponder・crop・Platform 分岐を集約し、
モーダルは UI 描画のみに整理する（287→134行）。
